third_party/WebKit/LayoutTests/external/wpt/css/css-variables/vars-border-shorthand-serialize-expected.txt

This is a testharness.js-based test.
FAIL border-color should serialize to the empty string when border references a variable assert_equals: expected "" but got "var(--border)"
FAIL border-style should serialize to the empty string when border references a variable assert_equals: expected "" but got "var(--border)"
FAIL border-width should serialize to the empty string when border references a variable assert_equals: expected "" but got "var(--border)"
Harness: the test ran to completion.
